refactor: unify user management with new User model
- Replaced AdminUser with User model for streamlined management
- Updated teacher creation & bulk upload logic for new schema
- Adjusted associations between User, AdminRole, and AdminSchool
- Refactored routes and controllers to align with new structure
- Improved error handling and response messages
- Enhanced S3 utility to support avatar uploads
- Cleaned up model definitions for consistency
- Removed deprecated AdminUser model and related references

+++ src/models/teacher_class_section_subject.model.js
@@ -6,7 +6,7 @@ export default sequelize.define("TeacherClassSectionSubject", {
     teacher_id: DataTypes.BIGINT,
     class_id: DataTypes.INTEGER,
     section_id: DataTypes.INTEGER,
-    class_subject_id: DataTypes.BIGINT,
+    class_subject_id: DataTypes.INTEGER,
     academic_year: DataTypes.STRING
 },{
     tableName: "teacher_class_section_subject",
